Final commit 9bal review madame Safa

+++ apps/hospitalisation/models.py
@@ -55,7 +55,7 @@ class Hospitalisation(models.Model):
         }
 
 class Lits_occupes(models.Model):
-    code_uf = models.IntegerField(db_column='code_uf', null=True, blank=True, help_text="Code de l'unité fonctionnelle")
+    code_uf = models.BigIntegerField(db_column='code_uf', null=True, blank=True, help_text="Code de l'unité fonctionnelle")
     value = models.IntegerField(null=True, blank=True, help_text="Nombre de lits occupés")
     date = models.DateField(null=True, blank=True, help_text="Date")
